requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
pydantic
python-jose
cachetools
alembic
pydantic_settings
pydantic[email]
